+++ src/pages/HomePage.tsx
@@ -1,24 +1,20 @@
 import React, { Component } from 'react'
-import Menu from 'UI/Menu/Menu'
 import { withStyles, Theme, Box, Typography, IconButton } from '@material-ui/core'
 import PageContent from 'UI/PageContent'
 import AddCircleIcon from '@material-ui/icons/AddCircle'
 import FilterListIcon from '@material-ui/icons/FilterList'
 import SearchInput from 'UI/SearchInput'
-import { connect } from 'react-redux'
+import PageWithMenu from 'UI/PageWithMenu'
 
 type HomePageProps = {
-    page: any,
-    onTabChange?(selected: string): void,
     classes: any
 }
 
 class HomePage extends Component<HomePageProps> {
     render() {
-        const { onTabChange, classes, page } = this.props
+        const { classes } = this.props
         return (
-            <Box className={classes.root}>
-                <Menu username="haspen" selectedTab={page} onTabChange={onTabChange}/>
+            <PageWithMenu>
                 <PageContent>
                     <Box className={classes.title}>
                         <Box>
@@ -35,21 +31,12 @@ class HomePage extends Component<HomePageProps> {
                         </Box>
                     </Box>
                 </PageContent>
-            </Box>
+            </PageWithMenu>
         )
     }
 }
 
-const mapStateToProps = ({ page }: any) => ({ page })
-const mapDispatchToProps = (dispatch: any) => ({ onTabChange: (selected: string) => dispatch({ type: selected }) })
-const HomePageWithState = connect(mapStateToProps, mapDispatchToProps)(HomePage)
-
 export default withStyles(({ palette }: Theme) => ({
-    root: {
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'left'
-    },
     title: {
         display: 'flex',
         width: '100%',
@@ -67,4 +54,4 @@ export default withStyles(({ palette }: Theme) => ({
         width: '30vw',
         minWidth: '200px',
     }
-}))(HomePageWithState)
+}))(HomePage)
